refactor: 기존의 CarFfeineMarker 구조 개선
[#811]
