+++ frontend/src/components/google-maps/marker/StationMarkersContainer/components/CarFfeineMarker/index.ts
@@ -1,0 +1,3 @@
+import CarFfeineMarker from './CarFfeineMarker';
+
+export default CarFfeineMarker;
