--- frontend/src/components/ui/CarFfeineMarker/index.ts
@@ -1,3 +1,0 @@
-import { CarFfeineMarker } from './CarFfeineMarker';
-
-export default CarFfeineMarker;
